+++ lang/en/block_panopto.php
@@ -34,11 +34,13 @@ $string['api_manager_unavailable'] = 'Unable to create the {$a} manager api clie
 $string['back_to_config'] = 'Back to config';
 $string['back_to_course'] = 'Back to course';
 $string['begin_reinitializing_imports'] = 'Begin Reinitializing Imports';
+$string['begin_folder_upgrade'] = 'Begin upgrading folders';
 $string['block_edit_error'] = 'Cannot configure block instance: Global configuration incomplete. Please contact your system administrator.';
 $string['block_edit_header'] = 'Select the Panopto course folder to display in this block.';
 $string['block_edit_header_help'] = 'Choose an existing Panopto course folder or provision a new Panopto course folder for this block';
 $string['block_global_add_courses'] = 'Add Moodle courses to Panopto';
 $string['block_global_reinitialize_all_imports'] = 'Reinitialize all Panopto folder imports';
+$string['block_global_upgrade_all_folders'] = 'Upgrade all existing folders and imports';
 $string['block_panopto_anyone_view_recorder_links'] = 'Allow all roles to view recorder download links';
 $string['block_panopto_anyone_view_recorder_links_desc'] = 'By default only users with creator/provisioning access on a folder can view the download links. If you wish viewers to be able to view the recorder download links through the block then enable this option.';
 $string['block_panopto_any_creator_can_view_folder_settings'] = 'Allow all users with creator roles to view Panopto folder settings links';
